refactor: hao.nguyen3 reviewd
