--- .env.stag
@@ -1,27 +1,0 @@
-# ENV NAME
-ENV=STAGING
-# APP NAME
-APP_NAME=Demo Staging
-
-#API URL
-API_URL=https://dev-api.carwash.sotatek.works
-
-# Android app ID. Define on google play
-ANDROID_APP_ID=com.demo.staging
-
-#Android app version name, define on google play
-ANDROID_APP_VERSION_NAME=1.0.0
-
-# Android version code, this is an unique code on google play
-ANDROID_APP_VERSION_CODE=1
-
-# Bundle ID of app on apple store
-IOS_APP_ID=com.demo.staging
-
-# Version name on apple store
-IOS_APP_VERSION_CODE=1.0.0
-
-# Build code on app store, unquire for each version
-IOS_APP_BUILD_CODE=1
-
-ENCRYPTION_KEY=a769504dead7ff32b18d34953510c97972d0ddf4bca6ab66c2818987136d05f5
