+++ package.json
@@ -55,7 +55,6 @@
         "react-native-safe-area-context": "^4.6.3",
         "react-native-screens": "^3.20.0",
         "react-native-svg": "^13.9.0",
-        "react-native-switch": "^1.5.1",
         "react-native-tab-view": "^3.5.1",
         "react-native-toast-notifications": "^3.3.1",
         "react-native-user-agent": "^2.3.1",
