--- src/components/index.ts
@@ -1,4 +1,0 @@
-import Input from 'components/Input';
-import FormInput from './FormInput';
-
-export { Input, FormInput };
