+++ src/packages/uikit/components/Button/index.ts
@@ -1,0 +1,3 @@
+export { default as Button } from './Button';
+export { default as ButtonIcon } from './ButtonIcon';
+export { default as ButtonText } from './ButtonText';
