+++ src/packages/uikit/components/Input/index.ts
@@ -1,0 +1,2 @@
+export { default as Input } from './Input';
+export { default as FormInput } from './FormInput';
