+++ src/packages/uikit/theme/base/radius.ts
@@ -1,3 +1,4 @@
+/* eslint-disable prettier/prettier */
 const radii = {
     none: 0,
     xs: 2,
@@ -5,8 +6,8 @@ const radii = {
     md: 6,
     lg: 8,
     xl: 12,
-    _2xl: 16,
-    _3xl: 24,
+    '2xl': 16,
+    '3xl': 24,
     full: 9999,
 };
